check for files before deleting on error
RunHostCmd tries to os.remove two file paths (rpc.REQ_FIFO and rpc.RES_FIFO)
in a finally block but those may not be created if an exception happens
early enough (for example, with an invalid api.idl file). This checks
for those files before calling os.remove on them, and silently catches
exceptions thrown when attempting to delete them.

A future improvement could also check for permissions before trying to
delete them (or create temporary files that are automatically cleaned up
instead if possible).

+++ stackhut_toolkit/commands.py
@@ -373,8 +373,18 @@ class RunHostCmd(HutCmd, UserCmd):
             log.error('Exception while running service: %s', str(err))
         finally:
             toolkit_stack.del_shim()
-            os.remove(rpc.REQ_FIFO)
-            os.remove(rpc.RESP_FIFO)
+
+            # TODO: remove nested try blocks
+            if os.path.exists(rpc.REQ_FIFO):
+                try:
+                    os.remove(rpc.REQ_FIFO)
+                except:
+                    pass
+            if os.path.exists(rpc.RESP_FIFO):
+                try:
+                    os.remove(rpc.RESP_FIFO)
+                except:
+                    pass
 
 
 class DeployCmd(HutCmd, UserCmd):
